add new publication with storing image to database

+++ src/main/java/com/kuzmenchuk/publications/repository/PublicationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PublicationRepository extends JpaRepository<Publication, Integer> {
-    Publication findPublicationById(Integer id);
+public interface PublicationRepository extends JpaRepository<Publication, Long> {
+    Publication findPublicationById(Long id);
 }
